Introduce the notion of a copy precondition token. (#14)
It's an opaque token that, when retrieved, can stop a copy if the key is overwritten with different data than we expect.  Given a bucket/key/checksum, a token is provided that can be passed to `copy(..)` and inhibits the copy if the token does not match what we're expecting.

On S3, it's just etag + CopySourceIfMatch.  On GS, it's generation id.